Added populating map for stops

+++ app/src/main/java/nz/zhang/aucklandtransportwear/atapi/Stop.kt
@@ -3,8 +3,8 @@ package nz.zhang.aucklandtransportwear.atapi
 class Stop (
         val stop_id: String,
         val stop_name: String,
-        val stop_lat: Float,
-        val stop_lon: Float,
+        val stop_lat: Double,
+        val stop_lon: Double,
         val stop_code: Int,
         val location_type: Int
 )
